fix vipcs lint errors

+++ includes/external-connection-cpt.php
@@ -164,13 +164,15 @@ function setup_list_table() {
 		if ( 'bulk-delete' === $doaction ) {
 			$sendback = remove_query_arg( array( 'trashed', 'untrashed', 'deleted', 'locked', 'ids' ), wp_get_referer() );
 
-			$deleted  = 0;
-			$post_ids = array_map( 'intval', $_REQUEST['post'] );
+			$deleted = 0;
+			if ( ! empty( $_REQUEST['post'] ) ) {
+				$post_ids = array_map( 'intval', $_REQUEST['post'] );
 
-			foreach ( (array) $post_ids as $post_id ) {
-				wp_delete_post( $post_id );
+				foreach ( (array) $post_ids as $post_id ) {
+					wp_delete_post( $post_id );
 
-				$deleted++;
+					$deleted ++;
+				}
 			}
 			$sendback = add_query_arg( 'deleted', $deleted, $sendback );
 
@@ -233,12 +235,8 @@ function ajax_verify_external_connection() {
 		exit;
 	}
 
-	$auth = array();
-	if ( ! empty( $_POST['auth'] ) ) {
-		$auth = $_POST['auth'];
-	}
-
-	$current_auth = get_post_meta( intval( sanitize_key( $_POST['endpointId'] ) ), 'dt_external_connection_auth', true );
+	$auth         = filter_input( INPUT_POST, 'auth', FILTER_REQUIRE_ARRAY );
+	$current_auth = get_post_meta( intval( sanitize_key( $_POST['endpointId'] ) ), 'dt_external_connection_auth', true ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotValidated
 
 	if ( ! empty( $current_auth ) ) {
 		$auth = array_merge( $auth, (array) $current_auth );
@@ -250,7 +248,7 @@ function ajax_verify_external_connection() {
 	$auth_handler = new $external_connection_class::$auth_handler_class( $auth );
 
 	// Init with placeholders since we haven't created yet
-	$external_connection = new $external_connection_class( 'connection-test', $_POST['url'], 0, $auth_handler );
+	$external_connection = new $external_connection_class( 'connection-test', esc_url( $_POST['url'] ), 0, $auth_handler );
 
 	$external_connections = $external_connection->check_connections();
 
@@ -355,7 +353,9 @@ function save_post( $post_id ) {
 		return;
 	}
 
-	if ( empty( $_POST['dt_external_connection_details'] ) || ! wp_verify_nonce( $_POST['dt_external_connection_details'], 'dt_external_connection_details_action' ) ) {
+	$nonce = filter_input( INPUT_POST, 'dt_external_connection_details', FILTER_SANITIZE_STRING );
+
+	if ( empty( $nonce ) || ! wp_verify_nonce( $nonce, 'dt_external_connection_details_action' ) ) {
 		return;
 	}
 
@@ -381,10 +381,8 @@ function save_post( $post_id ) {
 		// Create an instance of the connection to test connections
 		$external_connection_class = \Distributor\Connections::factory()->get_registered()[ sanitize_key( $_POST['dt_external_connection_type'] ) ];
 
-		$auth = array();
-		if ( ! empty( $_POST['dt_external_connection_auth'] ) ) {
-			$auth = $_POST['dt_external_connection_auth'];
-		}
+		$auth = filter_input( INPUT_POST, 'dt_external_connection_auth', FILTER_REQUIRE_ARRAY );
+		$url  = filter_input( INPUT_POST, 'dt_external_connection_url', FILTER_VALIDATE_URL );
 
 		$current_auth = get_post_meta( $post_id, 'dt_external_connection_auth', true );
 
@@ -394,7 +392,7 @@ function save_post( $post_id ) {
 
 		$auth_handler = new $external_connection_class::$auth_handler_class( $auth );
 
-		$external_connection = new $external_connection_class( get_the_title( $post_id ), $_POST['dt_external_connection_url'], $post_id, $auth_handler );
+		$external_connection = new $external_connection_class( get_the_title( $post_id ), $url, $post_id, $auth_handler );
 
 		$external_connections = $external_connection->check_connections();
 
@@ -402,7 +400,7 @@ function save_post( $post_id ) {
 		update_post_meta( $post_id, 'dt_external_connection_check_time', time() );
 	}
 
-	if ( ! empty( $_POST['dt_external_connection_auth'] ) ) {
+	if ( ! empty( $auth ) ) {
 		$current_auth = get_post_meta( $post_id, 'dt_external_connection_auth', true );
 		if ( empty( $current_auth ) ) {
 			$current_auth = array();
@@ -411,7 +409,7 @@ function save_post( $post_id ) {
 		$connection_class         = \Distributor\Connections::factory()->get_registered()[ sanitize_key( $_POST['dt_external_connection_type'] ) ];
 		$auth_handler_class_again = $connection_class::$auth_handler_class;
 
-		$auth_creds = $auth_handler_class_again::prepare_credentials( array_merge( (array) $current_auth, $_POST['dt_external_connection_auth'] ) );
+		$auth_creds = $auth_handler_class_again::prepare_credentials( array_merge( (array) $current_auth, $auth ) );
 
 		$auth_handler_class_again::store_credentials( $post_id, $auth_creds );
 	}
@@ -497,7 +495,7 @@ function meta_box_external_connection_details( $post ) {
 			<label for="dt_external_connection_type"><?php esc_html_e( 'Authentication Method', 'distributor' ); ?></label><br>
 			<select name="dt_external_connection_type" class="external-connection-type-field" id="dt_external_connection_type">
 				<?php foreach ( $registered_external_connection_types as $slug => $external_connection_class ) : ?>
-					<option <?php selected( $slug, $external_connection_type ); ?> value="<?php echo esc_attr( $slug ); ?>"><?php echo esc_attr( $external_connection_class::$label ); ?></option>
+					<option <?php selected( $slug, $external_connection_type ); ?> value="<?php echo esc_attr( $slug ); ?>"><?php echo esc_html( $external_connection_class::$label ); ?></option>
 				<?php endforeach; ?>
 			</select>
 		</div>
@@ -780,7 +778,7 @@ function filter_post_updated_messages( $messages ) {
  */
 function get_rest_url( $site_url ) {
 
-	$source = wp_remote_get( $site_url );
+	$source = wp_remote_get( $site_url ); // phpcs:ignore WordPressVIPMinimum.Functions.RestrictedFunctions.wp_remote_get_wp_remote_get
 
 	if ( is_wp_error( $source ) ) {
 		return $source;
@@ -814,7 +812,7 @@ function get_remote_distributor_info() {
 		exit;
 	}
 
-	$rest_url = get_rest_url( $_POST['url'] );
+	$rest_url = get_rest_url( esc_url( $_POST['url'] ) );
 
 	if ( is_wp_error( $rest_url ) ) {
 		wp_send_json_error( $rest_url );
@@ -826,7 +824,7 @@ function get_remote_distributor_info() {
 	if ( function_exists( 'vip_safe_wp_remote_get' ) && \Distributor\Utils\is_vip_com() ) {
 		$response = vip_safe_wp_remote_get( $route, false, 3, 3, 10 );
 	} else {
-		$response = wp_remote_get( $route, [ 'timeout' => 5 ] );
+		$response = wp_remote_get( $route, [ 'timeout' => 5 ] ); // phpcs:ignore WordPressVIPMinimum.Functions.RestrictedFunctions.wp_remote_get_wp_remote_get, WordPressVIPMinimum.Performance.RemoteRequestTimeout.timeout_timeout
 	}
 
 	$body = json_decode( wp_remote_retrieve_body( $response ), true );
